refactor: move submission information / scores tables into separate components
